requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
pandas-ta
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
python-telegram-bot==21.3
openai==1.35.3
apscheduler==3.10.4
newsapi-python==0.2.7
pydantic==2.7.4
pydantic-settings==2.3.4
